Add files via upload
Could anyone try to load the cvs file into PIG. Mine isn't working properly. This is the smallest file of the dataset that I could test with
Also, I have uploaded the converter as well. It works properly but its weird as an output. If someone could open it with Excel and check. I dont have Excel on my Mac. 

The python script takes the son file as the argument. If anyone wants to try other files.